requirements.txt
tzdata
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.9
django-colorfield==0.6.3
django-cors-headers==3.11.0
django-filter==21.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
openapi-codec==1.3.2
packaging==21.3
Pillow==9.0.0
psycopg2-binary==2.9.3
PyJWT==2.3.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
uritemplate==4.1.1
urllib3==1.26.8
vobject==0.9.6.1
whitenoise==5.3.0
cloudinary==1.29.0
django-cloudinary-storage==0.3.0
django-environ==0.4.5
django-on-heroku==1.1.2